refactor: remove config & prefixing
this was reducing code readability for very little gain
