+++ packages/react/src/hooks/index.ts
@@ -1,5 +1,3 @@
-export * from './useColorScheme';
-export * from './useConfig';
 export * from './useExternalClick';
 export * from './useForwardRef';
 export * from './usePopper';
