--- packages/react/src/hooks/useConfig.ts
@@ -1,4 +1,0 @@
-import { useContext } from 'react';
-import { Config, NDSContext } from '../config';
-
-export const useConfig = (): Config => useContext(NDSContext);
